+++ packages/flame/lib/src/game/mixins/game.dart
@@ -110,15 +110,28 @@ mixin Game on Loadable implements Projector {
       ''');
     }
     _gameRenderBox = gameRenderBox;
+
+    onAttach();
   }
+
+  /// Called when the game has been attached. This can be overriden
+  /// to add logic that requires the game to already be attached
+  /// to the widget tree.
+  void onAttach() {}
 
   /// Marks game as no longer attached to any Flutter widget tree.
   ///
   /// Should not be called manually.
   void detach() {
     _gameRenderBox = null;
-    _size = null;
+
+    onDetach();
   }
+
+  /// Called after the game has left the widget tree.
+  /// This can be overriden to add logic that requires the game
+  /// not be on the flutter widget tree anymore.
+  void onDetach() {}
 
   /// Converts a global coordinate (i.e. w.r.t. the app itself) to a local
   /// coordinate (i.e. w.r.t. he game widget).