+++ packages/flame_flare/CHANGELOG.md
@@ -1,6 +1,7 @@
-## [1.0.0-releasecandidate.14]
+## [Next]
 
-- Migrate to NS (to the extent that Flare allows us) & migrate to Flame 1.0
+- Migrate to Null-Safety
+- Migrate to Flame 1.0
 
 ## [0.1.0] - API change, fix positioning