+++ packages/flame_svg/CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
-## [Next]
+## [1.0.0-releasecandidate.15]
  - Takes priority as an argument on SvgComponent
+ - Simplify implementation of `renderPosition`
 
 ## [1.0.0-rc4]
  - Update to flutter_svg 0.22.0
@@ -12,7 +13,7 @@
 ## [1.0.0-rc2]
  - Updated to use flame v1.0.0-rc8
  - Migrate to null-safety
- 
+
 ## [1.0.0-rc1]
  - Updated to use flame v1.0.0